fix: broadcast defined in correct controller & updated routes
