+++ app/routes/dashboard.php
@@ -18,7 +18,6 @@ Route::prefix('dashboard')->middleware(['auth:user'])->group(function(){
   
   Route::controller(DashboardController::class)->group(function() {
     Route::get('', 'dashboard')->name('dashboard');
-    Route::match(['get', 'post'], '/broadcast', 'broadcast')->name('broadcast');
     Route::get('/leaderboard', 'leaderboard')->name('leaderboard');
   });
   
@@ -49,6 +48,7 @@ Route::prefix('dashboard')->middleware(['auth:user'])->group(function(){
 
   Route::controller(TeamController::class)->group(function() {
     Route::get('/teams', 'teams')->name('teams');
+    Route::match(['get', 'post'], '/broadcast', 'broadcast')->name('broadcast');
     Route::match(['get', 'post'], '/team/{id}/broadcast', 'broadcast')->name('broadcast-to-team');
     Route::get('/team/{id}/submissions', 'viewTeamSubmissions')->name('view-team-submissions');
     Route::match(['get', 'post'], '/team/{id}/delete', 'deleteTeam')->name('delete-team');
